requirements.txt
astroid==2.2.5
Django==2.1.7
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.3
isort==4.3.15
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
